+++ src/widgets/stack.rs
@@ -3,7 +3,7 @@
 use figures::units::UPx;
 use figures::{IntoSigned, Rect, Round, ScreenScale, Size};
 
-use crate::context::{AsEventContext, EventContext, GraphicsContext, LayoutContext};
+use crate::context::{AsEventContext, EventContext, GraphicsContext, LayoutContext, Trackable};
 use crate::styles::components::IntrinsicPadding;
 use crate::styles::FlexibleDimension;
 use crate::value::{Generation, IntoValue, Value};